+++ packages/share/android/src/main/java/io/flutter/plugins/share/MethodCallHandler.java
@@ -1,0 +1,36 @@
+// Copyright 2019 The Flutter Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+
+package io.flutter.plugins.share;
+
+
+import androidx.annotation.NonNull;
+import io.flutter.plugin.common.MethodCall;
+import io.flutter.plugin.common.MethodChannel;
+import java.util.Map;
+
+/** Handles the method channel for the plugin. */
+public class MethodCallHandler implements MethodChannel.MethodCallHandler {
+
+  private Share share;
+
+  /** Constructs the MethodChannelHandler */
+  public MethodCallHandler(@NonNull Share share) {
+    this.share = share;
+  }
+
+  @Override
+  public void onMethodCall(MethodCall call, MethodChannel.Result result) {
+    if (call.method.equals("share")) {
+      if (!(call.arguments instanceof Map)) {
+        throw new IllegalArgumentException("Map argument expected");
+      }
+      // Android does not support showing the share sheet at a particular point on screen.
+      share.share((String) call.argument("text"), (String) call.argument("subject"));
+      result.success(null);
+    } else {
+      result.notImplemented();
+    }
+  }
+}
